% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{getReports}
\alias{getReports}
\title{Query Your Usersnaps}
\usage{
getReports(...)
}
\arguments{
\item{...}{query parameters to include. There's no documentation, but known
parameters are (1) includeapikey, which should always be \code{"false"}; (2)
offset; (3) limit; and (4) search, which by default is
\code{'[{"type":"state","id":"open"}]'}, the default in the Usersnap GUI. You
can also specify a "project" to query against a project other than your
default.}
}
\value{
A list of report entries.
}
\description{
Query Your Usersnaps
}
